LPS-36216 Revert integration tests
