+++ portal-impl/test/integration/com/liferay/portal/lar/BasePortletDataHandlerTestCase.java
@@ -15,38 +15,27 @@
 package com.liferay.portal.lar;
 
 import com.liferay.portal.kernel.dao.orm.FinderCacheUtil;
-import com.liferay.portal.kernel.lar.ExportImportPathUtil;
 import com.liferay.portal.kernel.lar.ManifestSummary;
 import com.liferay.portal.kernel.lar.PortletDataContext;
 import com.liferay.portal.kernel.lar.PortletDataContextFactoryUtil;
 import com.liferay.portal.kernel.lar.PortletDataHandler;
 import com.liferay.portal.kernel.lar.PortletDataHandlerBoolean;
-import com.liferay.portal.kernel.lar.PortletDataHandlerKeys;
-import com.liferay.portal.kernel.lar.StagedModelType;
 import com.liferay.portal.kernel.transaction.Transactional;
 import com.liferay.portal.kernel.util.LongWrapper;
 import com.liferay.portal.kernel.util.Time;
 import com.liferay.portal.kernel.xml.Element;
-import com.liferay.portal.kernel.xml.ElementHandler;
-import com.liferay.portal.kernel.xml.ElementProcessor;
 import com.liferay.portal.kernel.xml.SAXReaderUtil;
-import com.liferay.portal.kernel.zip.ZipReader;
+import com.liferay.portal.kernel.zip.ZipWriter;
+import com.liferay.portal.kernel.zip.ZipWriterFactoryUtil;
 import com.liferay.portal.model.Group;
 import com.liferay.portal.service.GroupLocalServiceUtil;
 import com.liferay.portal.util.GroupTestUtil;
 import com.liferay.portlet.PortletPreferencesImpl;
 
-import java.io.StringReader;
-
 import java.util.Date;
 import java.util.HashMap;
-import java.util.HashSet;
-import java.util.Iterator;
 import java.util.LinkedHashMap;
 import java.util.Map;
-import java.util.Set;
-
-import org.apache.xerces.parsers.SAXParser;
 
 import org.junit.After;
 import org.junit.Assert;
@@ -54,8 +43,6 @@ import org.junit.Before;
 import org.junit.Test;
 
 import org.powermock.api.mockito.PowerMockito;
-
-import org.xml.sax.InputSource;
 
 /**
  * @author Zsolt Berentey
@@ -79,59 +66,6 @@ public abstract class BasePortletDataHandlerTestCase extends PowerMockito {
 
 	@Test
 	@Transactional
-	public void testDeletions() throws Exception {
-		initExport();
-
-		addStagedModels();
-
-		portletDataContext.setEndDate(getEndDate());
-
-		portletDataHandler.prepareManifestSummary(portletDataContext);
-
-		ManifestSummary manifestSummary =
-			portletDataContext.getManifestSummary();
-
-		Map<String, LongWrapper> modelAdditionCounters =
-			manifestSummary.getModelAdditionCounters();
-
-		Map<String, LongWrapper> initialModelDeletionCounters =
-			new HashMap<String, LongWrapper>(modelAdditionCounters);
-
-		removeUnsupportedDeletionSystemEventStagedModelTypes(
-			initialModelDeletionCounters);
-
-		modelAdditionCounters.clear();
-
-		portletDataHandler.exportData(
-			portletDataContext, portletId, new PortletPreferencesImpl());
-
-		deleteStagedModels();
-
-		portletDataContext.setEndDate(getEndDate());
-
-		portletDataHandler.prepareManifestSummary(portletDataContext);
-
-		Map<String, LongWrapper> modelDeletionCounters =
-			manifestSummary.getModelDeletionCounters();
-
-		removeUnsupportedDeletionSystemEventStagedModelTypes(
-			modelDeletionCounters);
-
-		checkCounters(initialModelDeletionCounters, modelDeletionCounters);
-
-		modelDeletionCounters.clear();
-
-		DeletionSystemEventExporter deletionSystemEventExporter =
-			new DeletionSystemEventExporter();
-
-		deletionSystemEventExporter.exportDeletionSystemEvents(
-			portletDataContext);
-
-		checkDeletions(initialModelDeletionCounters);
-	}
-
-	@Test
-	@Transactional
 	public void testPrepareManifestSummary() throws Exception {
 		initExport();
 
@@ -147,7 +81,7 @@ public abstract class BasePortletDataHandlerTestCase extends PowerMockito {
 		Map<String, LongWrapper> modelAdditionCounters =
 			manifestSummary.getModelAdditionCounters();
 
-		Map<String, LongWrapper> initialModelAdditionCounters =
+		Map<String, LongWrapper> expectedModelAdditionCounters =
 			new HashMap<String, LongWrapper>(modelAdditionCounters);
 
 		modelAdditionCounters.clear();
@@ -155,23 +89,7 @@ public abstract class BasePortletDataHandlerTestCase extends PowerMockito {
 		portletDataHandler.exportData(
 			portletDataContext, portletId, new PortletPreferencesImpl());
 
-		Set<String> classNames = initialModelAdditionCounters.keySet();
-
-		Iterator<String> iterator = classNames.iterator();
-
-		while (iterator.hasNext()) {
-			String className = iterator.next();
-
-			LongWrapper counter = initialModelAdditionCounters.get(className);
-
-			if (counter.getValue() == 0) {
-				iterator.remove();
-			}
-		}
-
-		checkCounters(
-			manifestSummary.getModelAdditionCounters(),
-			initialModelAdditionCounters);
+		checkManifestSummary(expectedModelAdditionCounters);
 	}
 
 	protected void addBooleanParameter(
@@ -191,98 +109,42 @@ public abstract class BasePortletDataHandlerTestCase extends PowerMockito {
 
 	protected abstract void addStagedModels() throws Exception;
 
-	protected void checkCounters(
-		Map<String, LongWrapper> expectedCounters,
-		Map<String, LongWrapper> actualCounters) {
-
-		for (String className : expectedCounters.keySet()) {
-			LongWrapper expectedCounter = expectedCounters.get(className);
-			LongWrapper actualCounter = actualCounters.get(className);
-
-			Assert.assertEquals(
-				className, expectedCounter.getValue(),
-				actualCounter.getValue());
-		}
-
-		Assert.assertEquals(expectedCounters.size(), actualCounters.size());
-	}
-
-	protected void checkDeletions(
-			Map<String, LongWrapper> expectedModelDeletionCounters)
-		throws Exception {
-
-		final Map<String, LongWrapper> actualModelDeletionCounters =
-			new HashMap<String, LongWrapper>();
-
-		SAXParser saxParser = new SAXParser();
-
-		ElementHandler elementHandler = new ElementHandler(
-			new ElementProcessor() {
-
-				@Override
-				public void processElement(Element element) {
-					StagedModelType stagedModelType =
-						new StagedModelType(
-							element.attributeValue("class-name"),
-							element.attributeValue("referrer-class-name"));
-
-					LongWrapper counter = actualModelDeletionCounters.get(
-						stagedModelType.toString());
-
-					if (counter != null) {
-						counter.increment();
-					}
-					else {
-						actualModelDeletionCounters.put(
-							stagedModelType.toString(), new LongWrapper(1));
-					}
-				}
-
-			},
-			new String[] {"deletion-system-event"});
-
-		saxParser.setContentHandler(elementHandler);
-
-		ZipReader zipReader = (ZipReader)portletDataContext.getZipWriter();
-
-		String deletionSystemEventsXML = zipReader.getEntryAsString(
-			ExportImportPathUtil.getRootPath(portletDataContext) +
-				"/deletion-system-events.xml");
-
-		Assert.assertNotNull(deletionSystemEventsXML);
-
-		saxParser.parse(
-			new InputSource(new StringReader(deletionSystemEventsXML)));
+	protected void checkManifestSummary(
+		Map<String, LongWrapper> expectedModelAdditionCounters) {
 
 		ManifestSummary manifestSummary =
 			portletDataContext.getManifestSummary();
 
-		Set<String> classNames = expectedModelDeletionCounters.keySet();
+		Map<String, LongWrapper> modelAdditionCounters =
+			manifestSummary.getModelAdditionCounters();
 
-		Iterator<String> iterator = classNames.iterator();
+		int expectedModelAdditionCountersSize =
+			expectedModelAdditionCounters.size();
 
-		while (iterator.hasNext()) {
-			String className = iterator.next();
+		for (String manifestSummaryKey :
+				expectedModelAdditionCounters.keySet()) {
 
-			LongWrapper counter = expectedModelDeletionCounters.get(className);
+			LongWrapper expectedModelAdditionCounter =
+				expectedModelAdditionCounters.get(manifestSummaryKey);
 
-			if (counter.getValue() == 0) {
-				iterator.remove();
+			if (expectedModelAdditionCounter.getValue() == 0) {
+				expectedModelAdditionCountersSize--;
+			}
+			else {
+				LongWrapper modelAdditionCounter = modelAdditionCounters.get(
+					manifestSummaryKey);
+
+				Assert.assertEquals(
+					expectedModelAdditionCounter.getValue(),
+					modelAdditionCounter.getValue());
 			}
 		}
 
-		checkCounters(
-			manifestSummary.getModelDeletionCounters(),
-			actualModelDeletionCounters);
-		checkCounters(
-			expectedModelDeletionCounters, actualModelDeletionCounters);
+		Assert.assertEquals(
+			modelAdditionCounters.size(), expectedModelAdditionCountersSize);
 	}
 
 	protected abstract PortletDataHandler createPortletDataHandler();
-
-	protected void deleteStagedModels() throws Exception {
-		portletDataHandler.deleteData(portletDataContext, portletId, null);
-	}
 
 	protected Date getEndDate() {
 		return new Date();
@@ -298,18 +160,14 @@ public abstract class BasePortletDataHandlerTestCase extends PowerMockito {
 		Map<String, String[]> parameterMap =
 			new LinkedHashMap<String, String[]>();
 
-		parameterMap.put(
-			PortletDataHandlerKeys.DELETIONS,
-			new String[] {Boolean.TRUE.toString()});
-
 		addParameters(parameterMap);
 
-		TestReaderWriter testReaderWriter = new TestReaderWriter();
+		zipWriter = ZipWriterFactoryUtil.getZipWriter();
 
 		portletDataContext =
 			PortletDataContextFactoryUtil.createExportPortletDataContext(
 				stagingGroup.getCompanyId(), stagingGroup.getGroupId(),
-				parameterMap, getStartDate(), getEndDate(), testReaderWriter);
+				parameterMap, getStartDate(), getEndDate(), zipWriter);
 
 		rootElement = SAXReaderUtil.createElement("root");
 
@@ -322,44 +180,12 @@ public abstract class BasePortletDataHandlerTestCase extends PowerMockito {
 			missingReferencesElement);
 	}
 
-	protected void removeUnsupportedDeletionSystemEventStagedModelTypes(
-		Map<String, LongWrapper> expectedModelDeletionCounters) {
-
-		Set<String> stagedModelTypes = new HashSet<String>();
-
-		for (Object deletionSystemEventStagedModelType :
-				portletDataHandler.getDeletionSystemEventStagedModelTypes()) {
-
-			if (deletionSystemEventStagedModelType instanceof Class) {
-				Class<?> clazz = (Class<?>)deletionSystemEventStagedModelType;
-
-				stagedModelTypes.add(clazz.getName());
-			}
-			else {
-				stagedModelTypes.add(
-					deletionSystemEventStagedModelType.toString());
-			}
-		}
-
-		Set<Map.Entry<String, LongWrapper>> set =
-			expectedModelDeletionCounters.entrySet();
-
-		Iterator<Map.Entry<String, LongWrapper>> iterator = set.iterator();
-
-		while (iterator.hasNext()) {
-			Map.Entry<String, LongWrapper> entry = iterator.next();
-
-			if (!stagedModelTypes.contains(entry.getKey())) {
-				iterator.remove();
-			}
-		}
-	}
-
 	protected Element missingReferencesElement;
 	protected PortletDataContext portletDataContext;
 	protected PortletDataHandler portletDataHandler;
 	protected String portletId;
 	protected Element rootElement;
 	protected Group stagingGroup;
+	protected ZipWriter zipWriter;
 
 }
