+++ portal-impl/test/integration/com/liferay/portal/lar/ExportImportHelperUtilTest.java
@@ -30,6 +30,7 @@ import com.liferay.portal.kernel.util.StringUtil;
 import com.liferay.portal.kernel.util.Time;
 import com.liferay.portal.kernel.xml.Element;
 import com.liferay.portal.kernel.xml.SAXReaderUtil;
+import com.liferay.portal.kernel.zip.ZipReader;
 import com.liferay.portal.kernel.zip.ZipWriter;
 import com.liferay.portal.kernel.zip.ZipWriterFactoryUtil;
 import com.liferay.portal.model.Group;
@@ -53,6 +54,7 @@ import com.liferay.portlet.documentlibrary.service.DLFileEntryLocalServiceUtil;
 import com.liferay.portlet.documentlibrary.util.DLAppTestUtil;
 import com.liferay.portlet.journal.util.JournalTestUtil;
 
+import java.io.File;
 import java.io.InputStream;
 
 import java.lang.reflect.Field;
@@ -492,5 +494,80 @@ public class ExportImportHelperUtilTest extends PowerMockito {
 	private PortletDataContext _portletDataContextImport;
 	private StagedModel _referrerStagedModel;
 	private Group _stagingGroup;
+
+	private class TestReaderWriter implements ZipReader, ZipWriter {
+
+		@Override
+		public void addEntry(String name, byte[] bytes) {
+			_binaryEntries.add(name);
+		}
+
+		@Override
+		public void addEntry(String name, InputStream inputStream) {
+			_binaryEntries.add(name);
+		}
+
+		@Override
+		public void addEntry(String name, String s) {
+			_entries.put(name, s);
+		}
+
+		@Override
+		public void addEntry(String name, StringBuilder sb) {
+			_entries.put(name, sb.toString());
+		}
+
+		@Override
+		public void close() {
+		}
+
+		@Override
+		public byte[] finish() {
+			return new byte[0];
+		}
+
+		public List<String> getBinaryEntries() {
+			return _binaryEntries;
+		}
+
+		@Override
+		public List<String> getEntries() {
+			return new ArrayList<String>(_entries.keySet());
+		}
+
+		@Override
+		public byte[] getEntryAsByteArray(String name) {
+			return null;
+		}
+
+		@Override
+		public InputStream getEntryAsInputStream(String name) {
+			return null;
+		}
+
+		@Override
+		public String getEntryAsString(String name) {
+			return _entries.get(name);
+		}
+
+		@Override
+		public List<String> getFolderEntries(String path) {
+			return null;
+		}
+
+		@Override
+		public File getFile() {
+			return null;
+		}
+
+		@Override
+		public String getPath() {
+			return StringPool.BLANK;
+		}
+
+		private List<String> _binaryEntries = new ArrayList<String>();
+		private Map<String, String> _entries = new HashMap<String, String>();
+
+	}
 
 }
