--- portal-impl/test/integration/com/liferay/portal/lar/TestReaderWriter.java
@@ -1,105 +1,0 @@
-/**
- * Copyright (c) 2000-2013 Liferay, Inc. All rights reserved.
- *
- * This library is free software; you can redistribute it and/or modify it under
- * the terms of the GNU Lesser General Public License as published by the Free
- * Software Foundation; either version 2.1 of the License, or (at your option)
- * any later version.
- *
- * This library is distributed in the hope that it will be useful, but WITHOUT
- * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
- * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
- * details.
- */
-
-package com.liferay.portal.lar;
-
-import com.liferay.portal.kernel.util.StringPool;
-import com.liferay.portal.kernel.zip.ZipReader;
-import com.liferay.portal.kernel.zip.ZipWriter;
-
-import java.io.File;
-import java.io.InputStream;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-
-/**
- * @author Zsolt Berentey
- */
-public class TestReaderWriter implements ZipReader, ZipWriter {
-
-	@Override
-	public void addEntry(String name, byte[] bytes) {
-		_binaryEntries.add(name);
-	}
-
-	@Override
-	public void addEntry(String name, InputStream inputStream) {
-		_binaryEntries.add(name);
-	}
-
-	@Override
-	public void addEntry(String name, String s) {
-		_entries.put(name, s);
-	}
-
-	@Override
-	public void addEntry(String name, StringBuilder sb) {
-		_entries.put(name, sb.toString());
-	}
-
-	@Override
-	public void close() {
-	}
-
-	@Override
-	public byte[] finish() {
-		return new byte[0];
-	}
-
-	public List<String> getBinaryEntries() {
-		return _binaryEntries;
-	}
-
-	@Override
-	public List<String> getEntries() {
-		return new ArrayList<String>(_entries.keySet());
-	}
-
-	@Override
-	public byte[] getEntryAsByteArray(String name) {
-		return null;
-	}
-
-	@Override
-	public InputStream getEntryAsInputStream(String name) {
-		return null;
-	}
-
-	@Override
-	public String getEntryAsString(String name) {
-		return _entries.get(name);
-	}
-
-	@Override
-	public File getFile() {
-		return null;
-	}
-
-	@Override
-	public List<String> getFolderEntries(String path) {
-		return null;
-	}
-
-	@Override
-	public String getPath() {
-		return StringPool.BLANK;
-	}
-
-	private List<String> _binaryEntries = new ArrayList<String>();
-	private Map<String, String> _entries = new HashMap<String, String>();
-
-}
